requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-telegram-bot==20.7
langchain==0.2.12
langchain-openai==0.1.20
faiss-cpu==1.8.0.post1
PyPDF2==3.0.1
beautifulsoup4==4.12.3
markdown==3.6
pydantic==2.9.2
python-dotenv==1.0.1